requirements.txt
pandas
langchain>=0.0.300
chromadb
openai
fastapi
uvicorn[standard]
python-dotenv
streamlit~=1.50.0
langchain-community~=0.3.31
langchain-core~=0.3.79
langchain-text-splitters~=0.3.11
typing_extensions~=4.15.0
requests~=2.32.5
bs4~=0.0.2
beautifulsoup4~=4.14.2
langgraph~=0.6.10
langchain_nvidia_ai_endpoints
langchain-astradb